refactor(headless): inline thin prompt ab wrappers
